# dist_api/swagger.json
{
	"basePath": "/v1",
	"definitions": {
		"User": {
			"description": "",
			"properties": {
				"id": {
					"type": "string",
					"description": ""
				},
				"firstName": {
					"type": "string",
					"description": ""
				},
				"lastName": {
					"type": "string",
					"description": ""
				},
				"age": {
					"type": "number",
					"format": "double",
					"description": ""
				}
			},
			"type": "object",
			"required": [
				"firstName",
				"lastName",
				"age"
			]
		}
	},
	"info": {
		"description": "a description",
		"license": {
			"name": "MIT"
		},
		"title": "Typescript-rest Test API",
		"version": "1.0"
	},
	"paths": {
		"/users": {
			"get": {
				"operationId": "UserServiceAllUsers",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "array",
							"items": {
								"type": "object"
							}
						}
					}
				},
				"description": "",
				"parameters": []
			},
			"post": {
				"operationId": "UserServiceCreateUser",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "object"
						}
					}
				},
				"description": "",
				"consumes": [
					"application/json"
				],
				"parameters": [
					{
						"description": "",
						"in": "body",
						"name": "user",
						"required": true,
						"schema": {
							"$ref": "#/definitions/User"
						}
					}
				]
			}
		},
		"/users/{id}": {
			"get": {
				"operationId": "UserServiceGetUser",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/User"
						}
					}
				},
				"description": "",
				"parameters": [
					{
						"description": "",
						"in": "path",
						"name": "id",
						"required": true,
						"type": "number",
						"format": "double"
					}
				]
			},
			"put": {
				"operationId": "UserServiceUpdateUser",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "object"
						}
					}
				},
				"description": "",
				"consumes": [
					"application/json"
				],
				"parameters": [
					{
						"description": "",
						"in": "path",
						"name": "id",
						"required": true,
						"type": "string"
					},
					{
						"description": "",
						"in": "body",
						"name": "user",
						"required": true,
						"schema": {
							"$ref": "#/definitions/User"
						}
					}
				]
			}
		}
	},
	"swagger": "2.0",
	"securityDefinitions": {
		"api_key": {
			"type": "apiKey",
			"name": "access_token",
			"in": "query"
		}
	},
	"host": "localhost:3000"
}